% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalParsing_fct.R
\name{axes_parser_binary}
\alias{axes_parser_binary}
\title{This function parses various elements of a graph definition
used to control the appearance of axes. They include
xlab, ylab, axes suppression of needed (fr ternary, mostly),
log scale.}
\usage{
axes_parser_binary(graphDef)
}
\arguments{
\item{graphDef}{A graph definition, loaded from json
/!\ uses GCDkit !! Should maybe move to Figaro}
}
\description{
This function parses various elements of a graph definition
used to control the appearance of axes. They include
xlab, ylab, axes suppression of needed (fr ternary, mostly),
log scale.
}
