% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalParsing_fct.R
\name{data_transformation}
\alias{data_transformation}
\title{This function transforms the data using the function required}
\usage{
data_transformation(graphDef, wrdata, transform_options)
}
\arguments{
\item{graphDef}{A graph definition, loaded from json}

\item{wrdata}{WR data (typically WR in GCDkit, or equivalent).
This is a matrix, as per GCDkit convention, but is actually
useful for ggplot as well as it will allow coordinate mapping for instance!
If used in ggplot, convert back to tibble.}

\item{transform_options}{Options to pass to the transformation function}
}
\description{
This function transforms the data using the function required
}
\details{
the transform fn may, or may not be GCDkit. Obviously if it is,
this depends on GCDkit...
}
