% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTransform_GCDkit.R
\name{DebonCalcFull}
\alias{DebonCalcFull}
\title{This function combines calculation of millications and of Debon parameters}
\usage{
DebonCalcFull(wrdata)
}
\arguments{
\item{wrdata}{A matrix with WR data}
}
\description{
This function combines calculation of millications and of Debon parameters
}
