% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTransform-fct.R
\name{demoTrans}
\alias{demoTrans}
\title{Showing how to use options}
\usage{
demoTrans(wrdata, doubleBB = F)
}
\arguments{
\item{wrdata}{a WR matrix}

\item{doubleBB}{a tag that will do something...}
}
\description{
Showing how to use options
}
