% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalParsing_fct.R
\name{.make.names}
\alias{.make.names}
\title{Ancillary function to get proper labels for axes, if supplied}
\usage{
.make.names(graphDef, W)
}
\arguments{
\item{graphDef}{A graph definition, loaded from json}

\item{W}{Axis to process
/!\ uses GCDkit !! Should maybe move to Figaro}
}
\description{
Ancillary function to get proper labels for axes, if supplied
}
