% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fct.R
\name{isColor}
\alias{isColor}
\title{Check whether a string (or strings) is a valid color name
https://stackoverflow.com/questions/13289009/check-if-character-string-is-a-valid-color-representation}
\usage{
isColor(x)
}
\arguments{
\item{x}{character vector, the strings to test}
}
\description{
Check whether a string (or strings) is a valid color name
https://stackoverflow.com/questions/13289009/check-if-character-string-is-a-valid-color-representation
}
