% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTransform-fct.R
\name{OhtaAraiParams}
\alias{OhtaAraiParams}
\title{Recalculate Otha + Arai M, F and W parameters}
\usage{
OhtaAraiParams(where)
}
\arguments{
\item{where}{A matrix with whole-rock composition, probably coming from GCDkit
#' /!\ uses GCDkit !! Should maybe move to Figaro}
}
\description{
Recalculate Otha + Arai M, F and W parameters
}
