% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalParsing_fct.R
\name{parse_template}
\alias{parse_template}
\title{Mostly, a wrapper to parse individual template elements
Works with template_element_parser}
\usage{
parse_template(graphDef, template_options = NULL, color_options = NULL)
}
\arguments{
\item{graphDef}{Graph definition list, loaded by json_loader}

\item{template_options}{List of all the options used in template, i.e. must match
template's "optionDefaults": field}

\item{color_options}{See plotDiagram_json}
}
\description{
Mostly, a wrapper to parse individual template elements
Works with template_element_parser
}
