% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Figaro_parser.R
\name{plotDiagram_json}
\alias{plotDiagram_json}
\title{Read a template defined in a json file and import it to  figaro graph}
\usage{
plotDiagram_json(
  json,
  path = NULL,
  wrdata = WR,
  lbl = labels,
  verbose = F,
  new = T,
  template_options = NULL,
  template_colors = NULL,
  transform_options = NULL
)
}
\arguments{
\item{json}{Name of the template file}

\item{path}{Path to json file
If not specified, looks in the json_template folder of the package}

\item{wrdata}{A matrix containing WR data. Probably GCDkit's WR,
as in many cases we need GCDkit-calculated things such as A/CNK etc.}

\item{lbl}{A data frame containing gCDkit labels.}

\item{verbose}{Give debugging info}

\item{new}{Open in a new window?}

\item{template_options}{Options passed to the parser, as named vector
They are of the form c("switch1" = T, "switch2" = F, etc)
and they will affect template elements that have a switch.}

\item{template_colors}{A named vector, with names corresponding to the colours
found in the template, e.g. c(col1="red",col2="blue"). Defaults to black.}

\item{transform_options}{Further options to be passed to data transformation function}
}
\description{
Read a template defined in a json file and import it to  figaro graph
}
\details{
This function reads a json template and plots a Figaro graph
 (figaro is GCDkit's internal plotting system, that allows
 a level of editing and interactivity). It can NOT work if GCDkit
 is not installed. This is the original function, intended as a
 replacement for GDkit::plotDiagram.

 This function will honour gCDkit global options, UNLESS the user overwrites them.

 Optional text (controlled by "switch": "showText" in the template): If the user specifies
 the option in the call, with template_options=c("showText"=T) for instance,
 the user choice will be followed. Else, the function will use options("gcd.plot.text").

 Likewise, if the user supplies colours for pltcol1, pltcol2 and pltcol3 they will be honoured.
 Otherwise, the system will use (1) black, if options("gcd.plot.bw") is true or
 (2) the content of plt.col[] if false.
}
