% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonReader.R
\name{plotDiagram_json_ternary}
\alias{plotDiagram_json_ternary}
\title{Inner function, for binary plots}
\usage{
plotDiagram_json_ternary(tpl, dd, lbl)
}
\arguments{
\item{tpl}{The template, loaded into a list from json}

\item{dd}{The plotting dataset}

\item{lbl}{The labels (GCDkit style)}
}
\description{
Inner function, for binary plots
}
\details{
Internal function, that does the actual plotting
in the case of a binary plot
}
