% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalParsing_fct.R
\name{points_coordinates}
\alias{points_coordinates}
\title{Calculate the x.data, y.data, a.data etc for any dataset, following template rules
This should include data transformation, and also filtering (if doFilter = T,
and of course if the template includes a filter).
This function should probably be made figaro-specific...#####
/!\ uses GCDkit !! Should maybe move to Figaro}
\usage{
points_coordinates(
  graphDef,
  wrdata,
  lbl,
  transform_options = NULL,
  doFilter = T
)
}
\arguments{
\item{graphDef}{A graph definition, loaded from json}

\item{wrdata}{WR data (typically WR in GCDkit, or equivalent)}

\item{lbl}{label data (typically labels in GCDkit, or equivalent)}

\item{doFilter}{Boolean, should the data be filtered according to template rule?}
}
\description{
Calculate the x.data, y.data, a.data etc for any dataset, following template rules
This should include data transformation, and also filtering (if doFilter = T,
and of course if the template includes a filter).
This function should probably be made figaro-specific...#####
/!\ uses GCDkit !! Should maybe move to Figaro
}
