% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTransform-fct.R
\name{recastAnhydrous}
\alias{recastAnhydrous}
\title{Recalculate WR composition to anhydrous, as done in gCDkit (First.r)
Ultimately there will probably be a GCDkit function by that name !}
\usage{
recastAnhydrous(where = WR)
}
\arguments{
\item{where}{A matrix with whole-rock composition, probably coming from GCDkit}
}
\description{
Recalculate WR composition to anhydrous, as done in gCDkit (First.r)
Ultimately there will probably be a GCDkit function by that name !
}
