% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalParsing_fct.R
\name{show_switches}
\alias{show_switches}
\title{Indicate the available options for this diagram}
\usage{
show_switches(graphDef)
}
\arguments{
\item{graphDef}{graph definition object, loaded from json}
}
\description{
Indicate the available options for this diagram
}
