% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalParsing_fct.R
\name{data_preparation}
\alias{data_preparation}
\title{This function prepares the data by filtering it,
if a filter condition was supplied; and by transforming it, if needed.}
\usage{
data_preparation(graphDef, wrdata, lbl)
}
\arguments{
\item{graphDef}{A graph definition, loaded from json}

\item{wrdata}{WR data (typically WR in GCDkit, or equivalent)}

\item{lbl}{label data (typically labels in GCDkit, or equivalent)}
}
\description{
This function prepares the data by filtering it,
if a filter condition was supplied; and by transforming it, if needed.
}
