% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalParsing_fct.R
\name{parse_template}
\alias{parse_template}
\title{Mostly, a wrapper to parse individual template elements
Works with template_element_parser}
\usage{
parse_template(graphDef, template_options = NULL, template_colors = NULL)
}
\arguments{
\item{template_options}{See plotDiagram_json}

\item{template_colors}{See plotDiagram_json}

\item{json}{Name of the template file}

\item{path}{Path to json file}
}
\description{
Mostly, a wrapper to parse individual template elements
Works with template_element_parser
}
