% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonReader.R
\name{plotDiagram_json}
\alias{plotDiagram_json}
\title{Read a template defined in a json file and import it to  figaro graph}
\usage{
plotDiagram_json(json, path = NULL, verbose = F)
}
\arguments{
\item{json}{Name of the template file}

\item{path}{Path to json file}

\item{verbose}{Give debugging info
If not specified, looks in the json_template folder of the package
Details
This function reads a json template and plots a Figaro graph
 (figaro is GCDkit's internal plotting system, that allows
 a level of editing and interactivity). It can NOT work if GCDkit
 is not installed. This is the original function, intended as a
 replacement for GDkit::plotDiagram.}
}
\description{
Read a template defined in a json file and import it to  figaro graph
}
