% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Figaro_parser.R
\name{plotDiagram_json_binary}
\alias{plotDiagram_json_binary}
\title{Inner function, for binary plots}
\usage{
plotDiagram_json_binary(
  graphDef,
  dd,
  lbl,
  new,
  template_options,
  template_colors
)
}
\arguments{
\item{graphDef}{The template, loaded into a list from json}

\item{dd}{The plotting dataset}

\item{lbl}{The labels (GCDkit style)}

\item{new}{Open in a new window?}

\item{template_options}{See plotDiagram_json}

\item{template_colors}{See plotDiagram_json}
}
\description{
Inner function, for binary plots
}
\details{
Internal function, that does the actual plotting
in the case of a binary plot
}
