% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalParsing_fct.R
\name{template_element_parser}
\alias{template_element_parser}
\title{This function looks at individual template elements,
and modifies them as desired}
\usage{
template_element_parser(tpl_el, template_options, template_colors)
}
\arguments{
\item{tpl_el}{A template element}
}
\description{
This function looks at individual template elements,
and modifies them as desired
}
